fix(video-metadata): make the audio credit form read as a form
The public-credit fields sit together inside one _InputWrapper card.
DivineTextField is transparent by default, so each field took the card's
colour and read as a label rather than an input — unlike the title and
description fields above, which each get their own card. Fill them so
they carry their own surface, split the "shared as" preview off with a
top border, and cross-fade the optional rows (provider credit vs public
editor, the source field, the validation line) instead of snapping the
card's height.

DivineSwitchTile and the cover screen's top bar pick up the same 16px
horizontal inset the rest of the metadata form uses.

Extract the section into its own file on the way. It carried three
classes and five imports that nothing else in video_metadata_form_fields
needed, and that file was over 500 lines.
